Refactored retrieve_filename's location
