+++ qhana_plugin_runner/api/util.py
@@ -326,51 +326,3 @@ class PluginUrl(ma.fields.URL):
         if value == "":
             value = None
         return super().deserialize(value, attr, data, **kwargs)
-
-
-# TODO find a more suitable place for _retrieve_filename and retrieve_filename. If this is found, then pull imports out of function.
-from requests.models import Response
-
-
-def _retrieve_filename(url: str, response: Response):
-    """
-    Given an url response and the original url to a file, it returns the name of the file
-    :param response: Response
-    :param url: str
-    :return: str
-    """
-    from pathlib import Path
-    from urllib.parse import urlparse
-
-    if "Content-Disposition" in response.headers.keys():
-        fname = re.findall("filename=(.+)", response.headers["Content-Disposition"])[0]
-        if fname[0] == fname[-1] and fname[0] in {'"', "'"}:
-            fname = fname[1:-1]
-    else:
-        fname = Path(urlparse(url).path).name
-    response.close()
-
-    # Remove file type endings
-    fname = Path(fname).stem
-
-    return fname
-
-
-def retrieve_filename(url_or_response: str | Response) -> str:
-    """
-    Given an url to a file or an url response, it returns the name of the file
-    :param url: str
-    :return: str
-    """
-    from ..requests import open_url
-
-    if isinstance(url_or_response, str):
-        url = url_or_response
-        with open_url(url) as response:
-            return _retrieve_filename(url, response)
-    elif isinstance(url_or_response, Response):
-        response = url_or_response
-        url = response.url
-        return _retrieve_filename(url, response)
-
-    raise ValueError("Expected input to be str or request.Response.")
