+++ stable_plugins/classical_ml/data_preparation/time_tanh.py
@@ -19,7 +19,6 @@ from json import dumps, loads
 from tempfile import SpooledTemporaryFile
 from typing import Mapping, Optional, List
 from zipfile import ZipFile
-import re
 
 import marshmallow as ma
 from celery.canvas import chain
@@ -45,7 +44,6 @@ from qhana_plugin_runner.api.util import (
     FrontendFormBaseSchema,
     SecurityBlueprint,
     FileUrl,
-    retrieve_filename,
 )
 from qhana_plugin_runner.celery import CELERY
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
@@ -53,7 +51,7 @@ from qhana_plugin_runner.plugin_utils.entity_marshalling import (
     save_entities,
     load_entities,
 )
-from qhana_plugin_runner.requests import open_url
+from qhana_plugin_runner.requests import open_url, retrieve_filename
 from qhana_plugin_runner.storage import STORE
 from qhana_plugin_runner.tasks import save_task_error, save_task_result
 from qhana_plugin_runner.util.plugins import QHAnaPluginBase, plugin_identifier
