+++ stable_plugins/classical_ml/data_preparation/transformers.py
@@ -40,7 +40,6 @@ from qhana_plugin_runner.api.plugin_schemas import (
     EntryPoint,
     DataMetadata,
     InputDataMetadata,
-    retrieve_filename,
 )
 from qhana_plugin_runner.api.util import (
     FrontendFormBaseSchema,
@@ -54,6 +53,7 @@ from qhana_plugin_runner.plugin_utils.zip_utils import get_files_from_zip_url
 from qhana_plugin_runner.storage import STORE
 from qhana_plugin_runner.tasks import save_task_error, save_task_result
 from qhana_plugin_runner.util.plugins import QHAnaPluginBase, plugin_identifier
+from qhana_plugin_runner.requests import retrieve_filename
 
 _plugin_name = "sim-to-dist-transformers"
 __version__ = "v0.2.0"
