+++ stable_plugins/classical_ml/scikit_ml/mds.py
@@ -15,7 +15,6 @@ from enum import Enum
 from http import HTTPStatus
 from tempfile import SpooledTemporaryFile
 from typing import Mapping, Optional
-import re
 
 import flask
 import marshmallow as ma
@@ -43,12 +42,11 @@ from qhana_plugin_runner.api.util import (
     FrontendFormBaseSchema,
     SecurityBlueprint,
     FileUrl,
-    retrieve_filename,
 )
 from qhana_plugin_runner.celery import CELERY
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
 from qhana_plugin_runner.plugin_utils.entity_marshalling import save_entities
-from qhana_plugin_runner.requests import open_url
+from qhana_plugin_runner.requests import open_url, retrieve_filename
 from qhana_plugin_runner.storage import STORE
 from qhana_plugin_runner.tasks import save_task_error, save_task_result
 from qhana_plugin_runner.util.plugins import QHAnaPluginBase, plugin_identifier
