+++ stable_plugins/classical_ml/scikit_ml/pca/tasks.py
@@ -27,9 +27,8 @@ from qhana_plugin_runner.plugin_utils.entity_marshalling import (
     load_entities,
     ensure_dict,
 )
-from qhana_plugin_runner.requests import open_url
+from qhana_plugin_runner.requests import open_url, retrieve_filename
 from qhana_plugin_runner.storage import STORE
-from qhana_plugin_runner.api.util import retrieve_filename
 
 from .pca_output import pca_to_output, get_output_dimensionality
