+++ stable_plugins/quantum_ml/max_cut/max_cut/tasks.py
@@ -31,7 +31,7 @@ from qhana_plugin_runner.plugin_utils.entity_marshalling import (
     save_entities,
 )
 from qhana_plugin_runner.storage import STORE
-from qhana_plugin_runner.api.util import retrieve_filename
+from qhana_plugin_runner.requests import retrieve_filename
 
 import numpy as np
