+++ stable_plugins/quantum_ml/pennylane_qiskit_ml/quantum_k_means/tasks.py
@@ -32,12 +32,11 @@ from qhana_plugin_runner.plugin_utils.entity_marshalling import (
     ensure_dict,
 )
 from qhana_plugin_runner.storage import STORE
-from qhana_plugin_runner.api.util import retrieve_filename
+from qhana_plugin_runner.requests import open_url, retrieve_filename
 
 import numpy as np
 
 from .backend.visualize import plot_data
-from qhana_plugin_runner.requests import open_url
 
 
 TASK_LOGGER = get_task_logger(__name__)
