+++ stable_plugins/visualization/complex/cluster_visualization.py
@@ -39,11 +39,10 @@ from qhana_plugin_runner.api.util import (
     FrontendFormBaseSchema,
     SecurityBlueprint,
     FileUrl,
-    retrieve_filename,
 )
 from qhana_plugin_runner.celery import CELERY
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
-from qhana_plugin_runner.requests import open_url
+from qhana_plugin_runner.requests import open_url, retrieve_filename
 from qhana_plugin_runner.storage import STORE
 from qhana_plugin_runner.tasks import save_task_error, save_task_result
 from qhana_plugin_runner.util.plugins import QHAnaPluginBase, plugin_identifier
